requirements.txt
pygame-ce
pygbag
pygame-gui